+++ src/std/tsf/index.test.ts
@@ -1,0 +1,51 @@
+import { tsf } from './index.js';
+
+tsf('');
+tsf('/');
+tsf('/')({});
+tsf('/{foo}')({ foo: (v) => v });
+tsf('/{foo}/{bar}')({ foo: 'hello', bar: () => 'world' });
+tsf('/{foo}/{bar}')({ foo: (v) => v, bar: (v) => v });
+tsf('/{foo}')({ foo: (v) => v.length > 1 && v.replace('o', 'u') });
+tsf('' as string)({ boo: (v) => v });
+tsf('' as `${string}/v1/posts/{id}/comments`)({ id: (v) => v });
+
+// ---- errors ----
+
+// @ts-expect-error
+tsf('{}');
+// @ts-expect-error
+tsf('/{}');
+// @ts-expect-error
+tsf('{}/');
+// @ts-expect-error
+tsf('/{{}}');
+// @ts-expect-error
+tsf('/{}}');
+// @ts-expect-error
+tsf('/{{}');
+// @ts-expect-error
+tsf('/{{}{');
+// @ts-expect-error
+tsf('/{{foo}}');
+// @ts-expect-error
+tsf('/{{foo}{');
+// @ts-expect-error
+tsf('/{}/{bar}');
+// @ts-expect-error
+tsf('/{foo}/{{}}');
+
+// @ts-expect-error
+tsf('/')();
+// @ts-expect-error
+tsf('/{foo}')();
+// @ts-expect-error
+tsf('/{foo}/{bar}')({});
+// @ts-expect-error
+tsf('/{foo}/{bar}')({ foo: (v) => v });
+// @ts-expect-error
+tsf('/{foo}')({ foo: (v) => v, bar: (v) => v });
+// @ts-expect-error
+tsf('' as `${string}/v1/posts/{id}/comments`)({});
+
+tsf('{hello-world}')({ 'hello-world': (v) => v });